+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Updated v4 API clients from v4 alpha to v4 beta  APIs
+- Handle trust bundle in v4 client cache GetOrCreate by setting VerifySSL
 
 ## [0.4.0] - 2024-05-03
 ### Added