+++ CHANGELOG.md
@@ -6,8 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+
+### Changed
+## [v0.5.0] - 2024-07-29
+### Added
 - Added v4 Categories beta APIs to v4 client
 - Added v4 VolumeGroups beta APIs to v4 client
+- Added a cache for v4 clients in v4 package
 
 ### Changed
 - Updated v4 API clients from v4 alpha to v4 beta  APIs